+++ python/cudf/cudf/tests/test_dataframe.py
@@ -7889,3 +7889,25 @@ def test_dataframe_to_pandas_nullable_dtypes(df, expected_pdf):
     actual_pdf = df.to_pandas(nullable=True)
 
     assert_eq(actual_pdf, expected_pdf)
+
+
+@pytest.mark.parametrize(
+    "op",
+    [
+        "count",
+        "cummin",
+        "cummax",
+        "cummax",
+        "cumprod",
+        "kurt",
+        "kurtosis",
+        "skew",
+    ],
+)
+def test_dataframe_axis1_unsupported_ops(op):
+    df = gd.DataFrame({"a": [1, 2, 3], "b": [8, 9, 10]})
+
+    with pytest.raises(
+        NotImplementedError, match="Only axis=0 is currently supported."
+    ):
+        getattr(df, op)(axis=1)